// src/app/interface/order-form.interface.ts
import { FormArray, FormControl, FormGroup } from '@angular/forms';
import { IOrderDetailForm } from './order-detail-form.interface';

export interface IOrderForm {
  name: FormControl<string | null>;
  address: FormControl<string | null>;
  phone: FormControl<string | null>;
  details: FormArray<FormGroup<IOrderDetailForm>>;
}